Code Restructuring
